+++ src/Modules/Database.js
@@ -12,6 +12,23 @@ async function CreatePendingAccount(Email, Password) {
     return UserRecord
 }
 
+async function CreateAccount(Email, Password, CreatedAt) {
+    let UserRecord = await prisma.accounts.create({
+        data: {
+            Email: Email,
+            Password: Password,
+            CreatedAt: CreatedAt
+        }
+    })
+
+    await prisma.profiles.create({
+        data: {
+            AccountId: UserRecord.id
+        }
+    })
+    return UserRecord
+}
+
 async function VerifyPendingAccount(PendingId) {
     let PendingRecord = await prisma.pendingAccounts.findUnique({
         where: {
@@ -19,40 +36,40 @@ async function VerifyPendingAccount(PendingId) {
         }
     })
     if (!PendingRecord) {
-        return false
+        return console.log("Account not found")
     }
-    let UserRecord = await prisma.accounts.create({
-        data: {
-            Email: PendingRecord.Email,
-            Password: PendingRecord.Password,
-            CreatedAt: PendingRecord.CreatedAt
-        }
-    })
+    console.log("Account found!")
+    await CreateAccount(PendingRecord.Email, PendingRecord.Password, PendingRecord.CreatedAt)
     await prisma.pendingAccounts.delete({
-        where: { id: PendingId }
-    })
-    return UserRecord
-}
-
-
-async function CreateAccount(Email, Password) {
-    let UserRecord = await prisma.accounts.create({
-        data: {
-            Email: Email,
-            Password: Password
+        where: {
+            id: PendingId
         }
+    }).then(() => {
+        console.log(`Deleted pending account with id ${PendingId}`)
     })
-    return UserRecord
 }
 
-async function UpdateAccount(Email, Password) {
+async function UpdateAccountEmail(Email) {
     let UserRecord = await prisma.accounts.update({
+        where: {
+            Email: Email
+        },
         data: {
-            Emaill: Email,
-            Password: Password
+            Email: Email
         }
     })
     return UserRecord
+}
+
+async function UpdateAccountPassword(userid, password) {
+    let UserRecord = await prisma.accounts.update({
+        where: {
+            id: userid
+        },
+        data: {
+            Password: password
+        }
+    })
 }
 
 async function FindAccountByEmail(Email) {
@@ -104,6 +121,7 @@ async function GetEmergencyInfo(AccountId) {
     let EmergencyData = await prisma.profiles.findMany({
         where: {
             id: AccountId
+
         },
         select: {
             Emergency_BloodType: true,
@@ -115,36 +133,64 @@ async function GetEmergencyInfo(AccountId) {
     return EmergencyData
 }
 
-/*
-async function CreateAccount(Email, Password) {
-    let UserRecord = await prisma.accounts.create({
-        data: {
-            Email: Email,
-            Password: Password
+async function FindProfile(id) {
+    let Profile = await prisma.profiles.findUnique({
+        where: {
+            AccountId: id
         }
     })
-    return UserRecord
+    if (Profile.Validated_ID == null && Profile.Unvalidate_ID == null){
+        delete(Profile.Validated_ID)
+        delete(Profile.Unvalidate_ID)
+        Profile["ID_Documents"] = ""
+    }
+    if (Profile.Validated_ID && Profile.Unvalidate_ID == null){
+        delete(Profile.Unvalidate_ID)
+        Profile["ID_Documents"] = Profile.Validated_ID
+    }
+    if (Profile.Validated_ID == null && Profile.Unvalidate_ID){
+        delete(Profile.Validated_ID)
+        Profile["ID_Documents"] = Profile.Unvalidate_ID
+    }
+    delete (Profile.id)
+    delete (Profile.AccountId)
+    return Profile
 }
-*/
 
-async function CreateProfile(AccountId, data) {
-    let ProfileData = await prisma.profiles.create({
-        data: {
-            Name: data.Name,
-            Gender: data.Gender,
-            School: data.School,
-            DOB: data.DOB,
-            ID_DOC: data.ID
+async function UpdateProfile(AccountId, data) {
+    let Profile = await FindProfile(AccountId)
+    let Remove = []
+    for (let key in Profile) {
+        if (!Profile[key] && !data[key]) {
+            Remove.push(key)
+            continue
         }
+        if (Profile[key] && data[key]) {
+            if (Profile[key] === data[key]) {
+                Remove.push(key)
+                continue
+            }
+            Profile[key] = data[key]
+        }
+        Profile[key] = data[key]
+    }
+    for (let i = 0; i < Remove.length; i++){
+        delete(Profile[Remove[i]])
+    }
+    console.log(Profile)
+    let NewProfile = await prisma.profiles.update({
+        where: {
+            AccountId: AccountId
+        },
+        data: Profile
     })
-}
-
-async function FindProfile(AccountId) {
-
-}
-
-async function UpdateProfile(AccountId) {
-
+    console.log("New Profile")
+    console.log(NewProfile)
+    delete (Profile.id)
+    delete(Profile.Validated_ID)
+    delete(Profile.Unvalidate_ID)
+    delete (Profile.AccountId)
+    return NewProfile
 }
 
 async function DeleteProfile(AccountId) {
@@ -155,9 +201,13 @@ async function DeleteProfile(AccountId) {
 module.exports = {
     CreateAccount: CreateAccount,
     CreatePendingAccount: CreatePendingAccount,
+    UpdateAccountEmail: UpdateAccountEmail,
+    UpdateAccountPassword: UpdateAccountPassword,
     VerifyPendingAccount: VerifyPendingAccount,
     FindAccountByEmail: FindAccountByEmail,
     FindPendingAccountByEmail: FindPendingAccountByEmail,
     FindAccountById: FindAccountById,
     GetEmergencyInfo: GetEmergencyInfo,
+    FindProfile: FindProfile,
+    UpdateProfile: UpdateProfile
 }