+++ src/Services/EmailService.js
@@ -8,11 +8,20 @@ const mailer = require("nodemailer")
 const { GenerateEmailToken } = require("./../Modules/Tokens")
 
 async function SendVerifyEmail(email, id){
-    let token = await GenerateEmailToken(email, id)
+    let token = await GenerateEmailToken(id)
+    //CLIENT_URL/verification/email/${token}
+    console.log(token)
+    return true
+}
+
+async function SendPasswordResetEmail(email, UserId, ip){
+    let token = await GeneratePasswordResetToken(UserId)
+    //CLIENT_URL/login/password-reset/${token}
     console.log(token)
     return true
 }
 
 module.exports = {
-    SendVerifyEmail: SendVerifyEmail
+    SendVerifyEmail: SendVerifyEmail,
+    SendPasswordResetEmail: SendPasswordResetEmail
 }