+++ src/config.js
@@ -11,6 +11,8 @@ module.exports = {
     MAX_REFRESH_TOKENS: process.env.MAX_REFRESH_TOKENS,
     EMAIL_TOKEN_SECRET: process.env.EMAIL_TOKEN_SECRET,
     EMAIL_TOKEN_EXPIRE: process.env.EMAIL_TOKEN_EXPIRE,
-
+    RESET_PASSWORD_SECRET: process.env.RESET_PASSWORD_SECRET,
+    RESET_PASSWORD_EXPIRE: process.env.RESET_PASSWORD_EXPIRE,
+    CLIENT_URL: process.env.CLIENT_URL,
     PORT: Number(process.env.PORT) || 8080
 }