+++ src/routes/profile.js
@@ -1,5 +1,7 @@
 const { AuthenticateAccessToken } = require("./../Middleware/AuthenticateToken")
-const {} = require("./../Modules/Database")
+const { FindProfile, UpdateProfile } = require("./../Modules/Database")
+const { ValidateDocuments, IsValidNumber, IsValidFacebookUrl } = require("./../Modules/Validate")
+
 const express = require("express")
 const crypto = require("crypto")
 const multer = require("multer")
@@ -12,14 +14,14 @@ const storage = multer.diskStorage({
         cb(null, path.join(__dirname, "..", "uploads", "images"))
     },
     filename: (req, file, cb) => {
-        cb (null, crypto.randomBytes(16).toString("hex") + path.extname(file.originalname))
+        cb(null, crypto.randomBytes(16).toString("hex") + path.extname(file.originalname))
     }
 })
 
 const upload = multer({
     storage: storage,
     fileFilter: (req, file, cb) => {
-        if (file.mimetype !== "image/jpg" && file.mimetype !== "image/jpeg" && file.mimetype != "image/png"){
+        if (file.mimetype !== "image/jpg" && file.mimetype !== "image/jpeg" && file.mimetype != "image/png") {
             cb(null, false)
             return cb(new Error("Only PNG, JPG and JPEG files are accepted!"))
         }
@@ -30,9 +32,49 @@ const upload = multer({
     }
 })
 
-router.post("/update", AuthenticateAccessToken,upload.single("image"), async (req, res) => {
-    
-    return res.status(200).json({message: "File uploaded!"})
+router.get("/", AuthenticateAccessToken, async (req, res) => {
+    let ip = req.headers["x-forwarded-for"] || req.socket.remoteAddress
+    let UserId = req.userid
+    let profile = await FindProfile(UserId)
+    return res.status(200).json({ success: true, profile })
+})
+
+router.post("/update", AuthenticateAccessToken, upload.single("image"), async (req, res) => {
+    let ip = req.headers["x-forwarded-for"] || req.socket.remoteAddress
+    let profile = JSON.parse(req.body.profile)
+    let UserId = req.userid
+    let doc_flag = false
+    if (profile.id || profile.AccountId) {
+        return res.status(403).json({ message: "Well...at least you tried" })
+    }
+    if (profile.ID_Documents){
+        let {success, doc} = ValidateDocuments(profile.ID_Documents)
+        doc_flag = success
+        profile.ID_Documents = doc
+    }
+    if (profile.Number){
+        if (!IsValidNumber(profile.Number)){
+            profile.Number = null
+        }
+    }
+    if (profile.Emergency_ContactNumber){
+        if (!IsValidNumber(profile.Emergency_ContactNumber)){
+            profile.Emergency_ContactNumber = null
+        }
+    }
+    if (profile.Emergency_BloodType){
+        if (!IsValidBloodType(profile.Emergency_BloodType)){
+            profile.Emergency_BloodType = null
+        }
+    }
+    if (profile.Facebook){
+        if (!IsValidFacebookUrl(profile.Facebook)){
+            profile.Facebook = null
+        }
+    }
+    let result = await UpdateProfile(UserId, profile)
+
+    return res.status(200).json({ message: "Profile updated" })
 })
 
 module.exports = router